refactor: make functions internal
