+++ src/data_factory_testing_framework/__init__.py
@@ -1,3 +1,6 @@
+from data_factory_testing_framework._functions.functions_repository import FunctionsRepository
+
 __all__ = [
     "test_framework",
+    "FunctionsRepository",
 ]
