+++ src/data_factory_testing_framework/_functions/evaluator/rules/dataset_reference_expression_rule_evaluator.py
@@ -1,20 +1,21 @@
 from lark import Tree
 
-from data_factory_testing_framework.functions.evaluator.exceptions import (
+from data_factory_testing_framework._functions.evaluator.exceptions import (
     ExpressionEvaluationInvalidChildTypeError,
     ExpressionEvaluationInvalidNumberOfChildrenError,
 )
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
+from data_factory_testing_framework._functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
+from data_factory_testing_framework.state.run_parameter_type import RunParameterType
 
 from .expression_rule_evaluator import ExpressionRuleEvaluator
 
 
-class ActivityReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
+class DatasetReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
     def __init__(self, tree: Tree, state: PipelineRunState) -> None:
         """Initialize expression rule evaluator."""
-        super().__init__(tree)
         self.state: PipelineRunState = state
+        super().__init__(tree)
 
         if len(self.children) != 1:
             raise ExpressionEvaluationInvalidNumberOfChildrenError(required=1, actual=len(self.children))
@@ -24,8 +25,11 @@ class ActivityReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
                 child_index=0, expected_types=EvaluationResult, actual_type=type(self.children[0])
             )
 
-        self.activity_name = self.children[0].value
+        self.parameter_name = self.children[0].value
 
     def evaluate(self) -> EvaluationResult:
-        activity_result = self.state.get_activity_result_by_name(self.activity_name)
-        return EvaluationResult(activity_result)
+        result = self.state.get_parameter_by_type_and_name(
+            RunParameterType.Dataset,
+            self.parameter_name,
+        )
+        return EvaluationResult(result)
