+++ src/data_factory_testing_framework/_functions/evaluator/rules/expression_rule_evaluator.py
@@ -4,7 +4,7 @@ from typing import Any, Union
 
 from lark import Tree
 
-from data_factory_testing_framework.functions.evaluator.exceptions import ExpressionEvaluationInvalidChildTypeError
+from data_factory_testing_framework._functions.evaluator.exceptions import ExpressionEvaluationInvalidChildTypeError
 
 
 @dataclass
