+++ src/data_factory_testing_framework/_functions/evaluator/rules/linked_service_reference_expression_rule_evaluator.py
@@ -1,19 +1,19 @@
 from lark import Tree
 
-from data_factory_testing_framework.functions.evaluator.exceptions import (
+from data_factory_testing_framework._functions.evaluator.exceptions import (
     ExpressionEvaluationInvalidChildTypeError,
     ExpressionEvaluationInvalidNumberOfChildrenError,
 )
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
+from data_factory_testing_framework._functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
 from data_factory_testing_framework.state.run_parameter_type import RunParameterType
 
 from .expression_rule_evaluator import ExpressionRuleEvaluator
 
 
-class SystemVariableReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
+class LinkedServiceReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
     def __init__(self, tree: Tree, state: PipelineRunState) -> None:
-        """Initialize the expression rule."""
+        """Initialize expression rule evaluator."""
         super().__init__(tree)
         self.state: PipelineRunState = state
 
@@ -25,11 +25,8 @@ class SystemVariableReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
                 child_index=0, expected_types=EvaluationResult, actual_type=type(self.children[0])
             )
 
-        self.system_variable_name = self.children[0].value
+        self.parameter_name = self.children[0].value
 
     def evaluate(self) -> EvaluationResult:
-        system_variable = self.state.get_parameter_by_type_and_name(
-            RunParameterType.System,
-            self.system_variable_name,
-        )
-        return EvaluationResult(system_variable)
+        parameter = self.state.get_parameter_by_type_and_name(RunParameterType.LinkedService, self.parameter_name)
+        return EvaluationResult(parameter)
