+++ src/data_factory_testing_framework/_functions/evaluator/rules/logical_bool_expression_rule_evaluator.py
@@ -2,12 +2,12 @@ from typing import Union
 
 from lark import Tree
 
-from data_factory_testing_framework.exceptions.expression_evaluation_error import ExpressionEvaluationError
-from data_factory_testing_framework.functions.evaluator.exceptions import (
+from data_factory_testing_framework._functions.evaluator.exceptions import (
     ExpressionEvaluationInvalidChildTypeError,
     ExpressionEvaluationInvalidNumberOfChildrenError,
 )
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
+from data_factory_testing_framework._functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
+from data_factory_testing_framework.exceptions.expression_evaluation_error import ExpressionEvaluationError
 
 from .expression_rule_evaluator import ExpressionRuleEvaluator
