+++ src/data_factory_testing_framework/_functions/evaluator/rules/system_variable_reference_expression_rule_evaluator.py
@@ -1,19 +1,19 @@
 from lark import Tree
 
-from data_factory_testing_framework.functions.evaluator.exceptions import (
+from data_factory_testing_framework._functions.evaluator.exceptions import (
     ExpressionEvaluationInvalidChildTypeError,
     ExpressionEvaluationInvalidNumberOfChildrenError,
 )
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
+from data_factory_testing_framework._functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
 from data_factory_testing_framework.state.run_parameter_type import RunParameterType
 
 from .expression_rule_evaluator import ExpressionRuleEvaluator
 
 
-class LinkedServiceReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
+class SystemVariableReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
     def __init__(self, tree: Tree, state: PipelineRunState) -> None:
-        """Initialize expression rule evaluator."""
+        """Initialize the expression rule."""
         super().__init__(tree)
         self.state: PipelineRunState = state
 
@@ -25,8 +25,11 @@ class LinkedServiceReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
                 child_index=0, expected_types=EvaluationResult, actual_type=type(self.children[0])
             )
 
-        self.parameter_name = self.children[0].value
+        self.system_variable_name = self.children[0].value
 
     def evaluate(self) -> EvaluationResult:
-        parameter = self.state.get_parameter_by_type_and_name(RunParameterType.LinkedService, self.parameter_name)
-        return EvaluationResult(parameter)
+        system_variable = self.state.get_parameter_by_type_and_name(
+            RunParameterType.System,
+            self.system_variable_name,
+        )
+        return EvaluationResult(system_variable)
