+++ src/data_factory_testing_framework/_functions/functions_repository.py
@@ -1,11 +1,11 @@
 from typing import Callable, Dict
 
-import data_factory_testing_framework.functions.functions_collection_implementation as collection_functions
-import data_factory_testing_framework.functions.functions_conversion_implementation as conversion_functions
-import data_factory_testing_framework.functions.functions_date_implementation as date_functions
-import data_factory_testing_framework.functions.functions_logical_implementation as logical_functions
-import data_factory_testing_framework.functions.functions_math_implementation as math_functions
-import data_factory_testing_framework.functions.functions_string_implementation as string_functions
+import data_factory_testing_framework._functions.functions_collection_implementation as collection_functions
+import data_factory_testing_framework._functions.functions_conversion_implementation as conversion_functions
+import data_factory_testing_framework._functions.functions_date_implementation as date_functions
+import data_factory_testing_framework._functions.functions_logical_implementation as logical_functions
+import data_factory_testing_framework._functions.functions_math_implementation as math_functions
+import data_factory_testing_framework._functions.functions_string_implementation as string_functions
 
 
 class FunctionsRepository:
