+++ tests/unit/functions/test_expression_evaluator.py
@@ -1,13 +1,13 @@
 from typing import Union
 
 import pytest
+from data_factory_testing_framework._functions.evaluator import ExpressionEvaluator
 from data_factory_testing_framework.exceptions.activity_not_found_error import ActivityNotFoundError
 from data_factory_testing_framework.exceptions.parameter_not_found_error import ParameterNotFoundError
 from data_factory_testing_framework.exceptions.state_iteration_item_not_set_error import (
     StateIterationItemNotSetError,
 )
 from data_factory_testing_framework.exceptions.variable_not_found_error import VariableNotFoundError
-from data_factory_testing_framework.functions import ExpressionEvaluator
 from data_factory_testing_framework.pythonnet.csharp_datetime import CSharpDateTime
 from data_factory_testing_framework.state.dependency_condition import DependencyCondition
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
