+++ tests/unit/functions/test_function_repository.py
@@ -1,6 +1,6 @@
 from unittest.mock import Mock
 
-from data_factory_testing_framework.functions.functions_repository import FunctionsRepository
+from data_factory_testing_framework._functions.functions_repository import FunctionsRepository
 
 
 def test_function_registration() -> None:
